solana_rpc_client_types

Module config

Structs

CommitmentConfig

RpcAccountInfoConfig

RpcBlockConfig

RpcBlockProductionConfig

RpcBlockProductionConfigRange

RpcBlockSubscribeConfig

RpcContextConfig

RpcEpochConfig

RpcGetVoteAccountsConfig

RpcLargestAccountsConfig

RpcLeaderScheduleConfig

RpcProgramAccountsConfig

RpcRequestAirdropConfig

RpcSendTransactionConfig

RpcSignatureStatusConfig

RpcSignatureSubscribeConfig

RpcSignaturesForAddressConfig

RpcSimulateTransactionAccountsConfig

RpcSimulateTransactionConfig

RpcSupplyConfig

RpcTransactionConfig

RpcTransactionLogsConfig

UiDataSliceConfig

Enums

CommitmentLevel

An attribute of a slot. It describes how finalized a block is at some point in time. For example, a slot is said to be at the max level immediately after the cluster recognizes the block at that slot as finalized. When querying the ledger state, use lower levels of commitment to report progress and higher levels to ensure state changes will not be rolled back.

RpcAccountIndex

RpcBlockSubscribeFilter

RpcBlocksConfigWrapper

RpcEncodingConfigWrapper

RpcLargestAccountsFilter

RpcLeaderScheduleConfigWrapper

RpcTokenAccountsFilter

RpcTransactionLogsFilter

TransactionDetails

UiAccountEncoding

UiTransactionEncoding

Traits

EncodingConfig